refactor: streamline domain validation in AllowedDomainsTab and admin router
- Replaced tldts library usage with a shared domain validation utility for consistency across components.
- Simplified domain validation logic to enhance clarity and maintainability.
- Added user feedback for invalid domain entries in AllowedDomainsTab, improving user experience.
- Updated chat export functionality in ChatInterface and EmbedHeader to notify users when there are no messages to export.
