+++ apps/web/src/components/admin/tabs/AllowedDomainsTab.tsx
@@ -23,7 +23,7 @@ import { Alert, AlertDescription } from "@/components/ui/alert";
 import { ConfirmationDialog } from "@/components/ui/confirmation-dialog";
 import { trpc } from "@/lib/trpc";
 import { toast } from "sonner";
-import { parse, getPublicSuffix } from "tldts";
+import { validateDomainForAllowlist } from "@/lib/domain-validation";
 
 export function AllowedDomainsTab() {
   const [newDomain, setNewDomain] = useState("");
@@ -78,99 +78,12 @@ export function AllowedDomainsTab() {
 
     const trimmedDomain = newDomain.trim().toLowerCase();
 
-    // Basic format validation
-    const domainRegex =
-      /^\.?[a-zA-Z0-9]([a-zA-Z0-9-]*[a-zA-Z0-9])?(\.[a-zA-Z0-9]([a-zA-Z0-9-]*[a-zA-Z0-9])?)*$/;
-    if (!domainRegex.test(trimmedDomain)) {
-      toast.error(
-        "Please enter a valid domain (e.g., .edu, stanford.edu, uni-bonn.de)",
-      );
-      return;
-    }
-
-    // Safe educational TLDs that are allowed (specific to educational institutions)
-    const safeEducationalTLDs = [
-      ".edu",
-      "edu", // US education
-      ".ac.uk",
-      "ac.uk", // UK academia
-      ".ac.in",
-      "ac.in", // India academia
-      ".edu.in",
-      "edu.in", // India education
-      ".ac.nz",
-      "ac.nz", // New Zealand academia
-      ".ac.za",
-      "ac.za", // South Africa academia
-      ".ac.jp",
-      "ac.jp", // Japan academia
-      ".ac.kr",
-      "ac.kr", // South Korea academia
-      ".ac.cn",
-      "ac.cn", // China academia
-      ".ac.il",
-      "ac.il", // Israel academia
-      ".edu.au",
-      "edu.au", // Australia education
-      ".edu.cn",
-      "edu.cn", // China education
-      ".edu.br",
-      "edu.br", // Brazil education
-      ".edu.mx",
-      "edu.mx", // Mexico education
-      ".edu.ar",
-      "edu.ar", // Argentina education
-      ".edu.co",
-      "edu.co", // Colombia education
-      ".edu.eg",
-      "edu.eg", // Egypt education
-      ".edu.pk",
-      "edu.pk", // Pakistan education
-      ".edu.sg",
-      "edu.sg", // Singapore education
-      ".edu.my",
-      "edu.my", // Malaysia education
-      ".edu.ph",
-      "edu.ph", // Philippines education
-    ];
-
-    // If it's a safe educational TLD, allow it immediately
-    if (safeEducationalTLDs.includes(trimmedDomain)) {
-      await addDomainMutation.mutateAsync({
-        domain: trimmedDomain,
-      });
-      return;
-    }
-
-    // Use tldts to parse the domain using the Public Suffix List (industry standard)
-    // Add a dummy prefix if the domain starts with a dot for proper parsing
-    const testDomain = trimmedDomain.startsWith(".")
-      ? `example${trimmedDomain}`
-      : trimmedDomain;
-
-    const parsed = parse(testDomain);
-    const publicSuffix = getPublicSuffix(testDomain);
-
-    // Check if what they entered is ONLY a public suffix (too broad)
-    const domainWithoutDot = trimmedDomain.startsWith(".")
-      ? trimmedDomain.slice(1)
-      : trimmedDomain;
-
-    // If the domain without dot equals the public suffix, it's just a TLD (too broad)
-    if (publicSuffix && domainWithoutDot === publicSuffix) {
-      toast.error("⚠️ Broad TLD detected", {
-        description: `Adding "${trimmedDomain}" will allow ALL emails from this domain type. For security, please add specific institutions instead (e.g., "stanford.edu" or "uni-bonn.de", not "${trimmedDomain}")`,
+    // Use shared validation utility (same logic as backend)
+    const validation = validateDomainForAllowlist(trimmedDomain);
+    if (!validation.valid) {
+      toast.error("⚠️ Invalid domain", {
+        description: validation.reason,
         duration: 8000,
-      });
-      return;
-    }
-
-    // If parsed domain is null or invalid, block it
-    if (!parsed.domain) {
-      toast.error("Invalid domain", {
-        description:
-          "Please enter a valid domain name or educational TLD pattern",
-        duration: 6000,
       });
       return;
     }
